# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "GitCredentialRecordInfo": {
        "properties": {
          "host_owner": {
            "title": "Host Owner",
            "type": "string"
          },
          "rejected_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejected At"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          }
        },
        "required": [
          "host_owner",
          "updated_at"
        ],
        "title": "GitCredentialRecordInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SecretRecordInfo": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          },
          "used_by": {
            "items": {
              "type": "string"
            },
            "title": "Used By",
            "type": "array"
          }
        },
        "required": [
          "name",
          "updated_at",
          "used_by"
        ],
        "title": "SecretRecordInfo",
        "type": "object"
      },
      "SecretValueRequest": {
        "properties": {
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "SecretValueRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Palmimo Portal",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/git-credentials/{host_owner}": {
      "delete": {
        "description": "Remove a registered git credential. A no-op if none exists for ``host_owner``.\n\nRaises:\n    PortalError: 422 ``invalid_host_owner`` if ``host_owner`` does not\n        normalize to ``host/owner`` shape.",
        "operationId": "delete_git_credential_api_v1_git_credentials__host_owner__delete",
        "parameters": [
          {
            "in": "path",
            "name": "host_owner",
            "required": true,
            "schema": {
              "title": "Host Owner",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Git Credential",
        "tags": [
          "secrets"
        ]
      },
      "put": {
        "description": "Register or overwrite a GitHub ``host/owner`` credential token (design doc 4.2). Never returned by any endpoint.\n\n``host_owner`` is normalized (lowercase host, trimmed) before storage --\nsee :func:`~palmimo_portal.core.secrets.normalize_host_owner` -- so\n``GitHub.com/Foo/`` and ``github.com/Foo`` land on the same entry.\n\nAlso clears :attr:`~palmimo_portal.ports.AppRecord.credential_rejected`\non every app scoped to ``host_owner`` -- a rewritten credential must\nresume periodic update checks (design doc 3.6), not stay skipped until\none happens to succeed. Skipped when ``apps.json`` is corrupt: writing\n``clear_credential_rejected``'s result back would replace the corrupt\nfile with a fresh empty ledger, destroying whatever ``POST\n/apps/reset`` was meant to recover instead of the credential save.\n\nRaises:\n    PortalError: 422 ``invalid_host_owner`` if ``host_owner`` does not\n        normalize to ``host/owner`` shape; 422 ``invalid_secret_value``\n        if the token contains a newline.",
        "operationId": "put_git_credential_api_v1_git_credentials__host_owner__put",
        "parameters": [
          {
            "in": "path",
            "name": "host_owner",
            "required": true,
            "schema": {
              "title": "Host Owner",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretValueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitCredentialRecordInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Git Credential",
        "tags": [
          "secrets"
        ]
      }
    },
...
    "/api/v1/secrets/{name}": {
      "delete": {
        "description": "Delete a registered secret.\n\nRaises:\n    PortalError: 404 ``secret_not_found``; 409 ``secret_in_use`` (with\n        the apps/request-names still bound to it) if a binding\n        references it.",
        "operationId": "delete_secret_api_v1_secrets__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Secret",
        "tags": [
          "secrets"
        ]
      },
      "put": {
        "description": "Register or overwrite a secret value.\n\nRaises:\n    PortalError: 422 ``invalid_secret_name`` if ``name`` does not match\n        ``^[A-Z][A-Z0-9_]{0,63}$``; 422 ``invalid_secret_value`` if the\n        value contains a newline.",
        "operationId": "put_secret_api_v1_secrets__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SecretValueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretRecordInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Secret",
        "tags": [
          "secrets"
        ]
      }
    },
...
  }
}